+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Panic instead of asserting where applicable
 - Restructured code
 - Rename `lazyOnce()` and `lazyRepeat()` to `onceWith()` and `repeatWith()` respectively
+- `assertIsIter()` now returns an error union
 
 ### Removed